Change gameid to prevent overlap in MP with upstream
Games are segregated on version number and gameid.
The gameid consists in 4 bytes characters describing Hellfire, Hellfire
Spawn, Diablo or Diablo Spawn.

To get a segregation between this mod and upstream and keep the internal
segregation between Hellfire, Hellfire Spawn etc.. this patch does a
bitwise OR between the original 4 bytes and "IRON".