Allow descending in MP when another player has already descended
Because of desyncs and/or bugs, in Multi-Player there can be situations
where the counters are off for one player but not the others. The player
is then unable to descend.

In an ironman setting, when such case happens, it can be logically
assuemd that the level was cleared (otherwise, the other player would
not have been able to descend).

This patch implements this logic by checking whether any player is
already in the dlvl about to be entered.

Implements #27