Add objects highlighting
Finding the last chest in a level could be tedious, as the chest might
have been hidden near a wall or a door. This was forcing the player to
"spray" the level with their mouse to find the chest.

This patch adds objects highlighting, customizable through an option
with four possible values:
  - WhenSafe: highlights objects once the level is safe, i.e. when all
    monsters are killed. This is the default.
  - WhenAlt: highlights objects when Alt key is pressed
  - WhenAltAndSafe: combination of the two above. Alt key must be
    pressed, and all monsters must be killed
  - Never: vanilla behaviour, no object highlighting

Implements #29